--- tests/artifacts/tsMocks.ts
@@ -1,29 +1,0 @@
-
-
-
-
-
-export const tsExampleDec = {
-    "$schema": "https://raw.githubusercontent.com/F5Networks/f5-telemetry-streaming/master/src/schema/latest/base_schema.json",
-    "class": "Telemetry",
-    "My_System": {
-        "class": "Telemetry_System",
-        "systemPoller": {
-            "interval": 60
-        }
-    },
-    "My_Listener": {
-        "class": "Telemetry_Listener",
-        "port": 6514
-    },
-    "My_Consumer": {
-        "class": "Telemetry_Consumer",
-        "type": "Splunk",
-        "host": "192.0.2.1",
-        "protocol": "https",
-        "port": 8088,
-        "passphrase": {
-            "cipherText": "apikey"
-        }
-    }
-}
